backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic
openai
langchain
langchain-openai
langchain-community
chromadb
pypdf2
pytesseract
pillow
pdf2image
numpy
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]